perf(opencv、image): 优化代码

+++ commons-opencv/src/main/java/io/github/pangju666/commons/opencv/utils/OpenCvUtils.java
@@ -492,7 +492,7 @@ public class OpenCvUtils {
 	 * @return 2x3 的平移变换矩阵
 	 * @since 2.1.0
 	 */
-	public static Mat createAffineTranslateMatrix(final int dx, final int dy) {
+	public static Mat createAffineTranslateMatrix(final double dx, final double dy) {
 		Mat matrixMat = new Mat(2, 3, opencv_core.CV_64F);
 		try (DoubleIndexer indexer = matrixMat.createIndexer()) {
 			// 第一行 [1, 0, dx]
